ordertype分页

+++ xiaobing-service/src/main/java/com/zpy/xiaobingservice/service/OrdertypeService.java
@@ -2,8 +2,10 @@ package com.zpy.xiaobingservice.service;
 
 import com.zpy.xiaobingservice.entity.Order;
 import com.zpy.xiaobingservice.entity.Ordertype;
+import com.zpy.xiaobingservice.mapper.OrderMapper;
 import com.zpy.xiaobingservice.mapper.OrdertypeMapper;
 import com.zpy.xiaobingservice.qo.OrderTypeQo;
+import jdk.nashorn.internal.ir.IfNode;
 import org.springframework.stereotype.Service;
 
 import javax.annotation.Resource;
@@ -14,6 +16,9 @@ public class OrdertypeService {
 
     @Resource
     private OrdertypeMapper ordertypeMapper;
+    @Resource
+    private OrderMapper orderMapper;
+
     public Ordertype update(Ordertype ordertype) {
         ordertypeMapper.updateByPrimaryKey(ordertype);
         return ordertype;
@@ -21,10 +26,14 @@ public class OrdertypeService {
 
     public Ordertype insert(Ordertype ordertype) {
         ordertypeMapper.insert(ordertype);
-        return ordertype;
+        return ordertypeMapper.selectByPrimaryKey(ordertype.getId());
     }
 
     public int delete(Integer ordertypeId) {
+        int count = orderMapper.selectByOrderTypeId(ordertypeId);
+        if (count > 0) {
+            return -1;
+        }
         int row = ordertypeMapper.deleteByPrimaryKey(ordertypeId);
         return row;
     }
